refactor: extract time interval constants
