+++ apps/dapp/package.json
@@ -56,6 +56,7 @@
     "react-select": "^4.3.1",
     "react-typical": "^0.1.3",
     "react-vis": "^1.11.7",
+    "recharts": "^2.4.2",
     "styled-components": "^5.3.3",
     "tippy.js": "^6.3.7",
     "use-debounce": "^9.0.2",
