+++ apps/dapp/src/components/Pages/Analytics/index.tsx
@@ -1,0 +1,4 @@
+import type { FC } from 'react';
+import { TIME_INTERVAL, DEFAULT_CHART_INTERVALS } from 'utils/time-intervals';
+
+export const AnalyticsPage: FC = () => <div></div>;
